chore: Move routing definition files and minor refactoring
to make sure routing definition files is defined routes, and Refactor indents to format equal points
