+++ server/src/main/scala/com/example/tryakkahttpscalajs/routes/WebServer.scala
@@ -1,4 +1,4 @@
-package com.example.tryakkahttpscalajs
+package com.example.tryakkahttpscalajs.routes
 
 import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
@@ -7,12 +7,12 @@ import com.typesafe.config.ConfigFactory
 
 object WebServer {
   def main(args: Array[String]) {
-    implicit val system = ActorSystem("server-system")
-    implicit val materializer = ActorMaterializer()
+    implicit val system: ActorSystem             = ActorSystem("server-system")
+    implicit val materializer: ActorMaterializer = ActorMaterializer()
 
-    val config = ConfigFactory.load()
+    val config    = ConfigFactory.load()
     val interface = config.getString("http.interface")
-    val port = config.getInt("http.port")
+    val port      = config.getInt("http.port")
 
     val service = new WebService()
