+++ server/src/main/scala/com/example/tryakkahttpscalajs/twirl/Implicits.scala
@@ -8,9 +8,9 @@ import play.twirl.api.{Html, Txt, Xml}
 object Implicits {
 
   /** Twirl marshallers for Xml, Html and Txt mediatypes */
-  implicit val twirlHtmlMarshaller = twirlMarshaller[Html](`text/html`)
-  implicit val twirlTxtMarshaller  = twirlMarshaller[Txt](`text/plain`)
-  implicit val twirlXmlMarshaller  = twirlMarshaller[Xml](`text/xml`)
+  implicit val twirlHtmlMarshaller: ToEntityMarshaller[Html] = twirlMarshaller[Html](`text/html`)
+  implicit val twirlTxtMarshaller: ToEntityMarshaller[Txt]   = twirlMarshaller[Txt](`text/plain`)
+  implicit val twirlXmlMarshaller: ToEntityMarshaller[Xml]   = twirlMarshaller[Xml](`text/xml`)
 
   def twirlMarshaller[A](contentType: MediaType): ToEntityMarshaller[A] = {
     Marshaller.StringMarshaller.wrap(contentType)(_.toString)
